rename test and fix docker config(port)
