+++ cluster-game-server/test/application/connect/interactor_test.go
@@ -1,4 +1,4 @@
-package application_connect_test
+package interactor_test
 
 import (
 	"testing"
